added two casting which remove the linter warnings

+++ MIHCrypto/AES/MIHAESKey.m
@@ -138,7 +138,7 @@
             if (error)
                 *error = [NSError errorWithDomain:MIHCryptoErrorDomain
                                              code:MIHCryptoInvalidKeySize
-                                         userInfo:@{NSLocalizedDescriptionKey : [NSString stringWithFormat:@"%lu is not a valid AES key size!", self.key.length]}];
+                                         userInfo:@{NSLocalizedDescriptionKey : [NSString stringWithFormat:@"%lu is not a valid AES key size!", (unsigned long) self.key.length]}];
             free(messageBytes);
             return nil;
     }
